improve product brief and setup makefile

+++ setup.mk
@@ -1,6 +1,6 @@
 # (c) 2022-Present IObundle, Lda, all rights reserved
 #
-# This makefile is used to setup a build directory for an IP core
+# This file is run as a makefile to setup a build directory for an IP core
 #
 
 SHELL=/bin/bash
@@ -29,22 +29,11 @@ FPGA_TOOL=$(shell find $(LIB_DIR)/hardware/boards -name $(BOARD) | cut -d"/" -f5
 
 # establish build dir paths
 VERSION_STR := $(shell $(LIB_DIR)/software/python/version.py -i .)
-
 BUILD_DIR := $(NAME)_$(VERSION_STR)
-BUILD_SW_DIR:=$(BUILD_DIR)/sw
-BUILD_SW_PYTHON_DIR:=$(BUILD_SW_DIR)/python
-BUILD_SW_SRC_DIR:=$(BUILD_DIR)/sw/src
-BUILD_SW_BSRC_DIR:=$(BUILD_DIR)/sw/bsrc
-BUILD_SW_PC_DIR:=$(BUILD_SW_DIR)/pc
-BUILD_SW_PCSRC_DIR:=$(BUILD_DIR)/sw/pcsrc
-BUILD_SW_EMB_DIR:=$(BUILD_SW_DIR)/emb
 BUILD_VSRC_DIR:=$(BUILD_DIR)/hw/vsrc
 BUILD_SIM_DIR:=$(BUILD_DIR)/hw/sim
 BUILD_FPGA_DIR:=$(BUILD_DIR)/hw/fpga
 BUILD_DOC_DIR:=$(BUILD_DIR)/doc
-BUILD_TSRC_DIR:=$(BUILD_DOC_DIR)/tsrc
-BUILD_FIG_DIR:=$(BUILD_DOC_DIR)/figures
-BUILD_SYN_DIR:=$(BUILD_DIR)/hw/syn
 
 SRC+=$(BUILD_DIR)/info.mk
 $(BUILD_DIR)/info.mk:
@@ -63,6 +52,7 @@ SRC+=$(BUILD_DIR)/sw/pc
 SRC+=$(BUILD_DIR)/sw/pcsrc
 SRC+=$(BUILD_DIR)/sw/python
 SRC+=$(BUILD_DIR)/sw/src
+
 # create BUILD_DIR/sw directories
 $(BUILD_DIR)/sw $(BUILD_DIR)/sw/bash $(BUILD_DIR)/sw/bsrc $(BUILD_DIR)/sw/emb $(BUILD_DIR)/sw/pc $(BUILD_DIR)/sw/pcsrc $(BUILD_DIR)/sw/python $(BUILD_DIR)/sw/src:
 	mkdir -p $@
@@ -72,18 +62,19 @@ SRC+=$(BUILD_DIR)/sw/pc/Makefile
 $(BUILD_DIR)/sw/%/Makefile: $(LIB_DIR)/software/%/Makefile
 	cp $< $@
 
-SRC+=$(patsubst $(PC_DIR)/%, $(BUILD_SW_PC_DIR)/%, $(wildcard $(PC_DIR)/*.expected))
-$(BUILD_SW_PC_DIR)/%.expected: $(PC_DIR)/%.expected
+SRC+=$(patsubst $(PC_DIR)/%, $(BUILD_DIR)/sw/pc/%, $(wildcard $(PC_DIR)/*.expected))
+$(BUILD_DIR)/sw/pc/%.expected: $(PC_DIR)/%.expected
 	cp $< $@
 
-SRC+=$(patsubst $(PC_DIR)/%, $(BUILD_SW_PC_DIR)/%, $(wildcard $(PC_DIR)/pc-emul.mk))
-$(BUILD_SW_PC_DIR)/pc-emul.mk: $(PC_DIR)/pc-emul.mk
+SRC+=$(patsubst $(PC_DIR)/%, $(BUILD_DIR)/sw/pc/%, $(wildcard $(PC_DIR)/pc-emul.mk))
+$(BUILD_DIR)/sw/pc/pc-emul.mk: $(PC_DIR)/pc-emul.mk
 	cp $< $@
 
-SRC+=$(patsubst $(EMB_DIR)/%, $(BUILD_SW_EMB_DIR)/%, $(wildcard $(EMB_DIR)/embedded.mk))
-$(BUILD_SW_PC_DIR)/embedded.mk: $(EMB_DIR)/embedded.mk
+SRC+=$(patsubst $(EMB_DIR)/%, $(BUILD_DIR)/sw/emb/%, $(wildcard $(EMB_DIR)/embedded.mk))
+$(BUILD_DIR)/sw/pc/embedded.mk: $(EMB_DIR)/embedded.mk
 	cp $< $@
 endif
+
 #--------------------- SIMULATION-----------------------
 ifneq ($(wildcard hardware/simulation/.),)
 SRC+=$(BUILD_SIM_DIR)
@@ -110,6 +101,7 @@ SRC+=$(patsubst $(SIM_DIR)/%, $(BUILD_SIM_DIR)/%, $(wildcard $(SIM_DIR)/*.v))
 $(BUILD_SIM_DIR)/%.v: $(SIM_DIR)/%.v
 	cp $< $@
 endif
+
 #--------------------- FPGA-----------------------
 ifneq ($(wildcard hardware/fpga/.),)
 SRC+=$(BUILD_FPGA_DIR)
@@ -128,8 +120,8 @@ SRC+=$(patsubst $(LIB_DIR)/hardware/boards/$(FPGA_TOOL)/%, $(BUILD_FPGA_DIR)/fpg
 $(BUILD_FPGA_DIR)/fpga_tool.tcl: $(LIB_DIR)/hardware/boards/$(FPGA_TOOL)/$(FPGA_TOOL).tcl
 	cp $< $@
 
-SRC+=$(patsubst $(LIB_DIR)/software/bash/%, $(BUILD_SW_DIR)/bash/%, $(wildcard $(LIB_DIR)/software/bash/$(FPGA_TOOL)2tex.sh))
-$(BUILD_SW_DIR)/bash/%: $(LIB_DIR)/software/bash/%
+SRC+=$(patsubst $(LIB_DIR)/software/bash/%, $(BUILD_DIR)/sw/bash/%, $(wildcard $(LIB_DIR)/software/bash/$(FPGA_TOOL)2tex.sh))
+$(BUILD_DIR)/sw/bash/%: $(LIB_DIR)/software/bash/%
 	cp $< $@
 
 SRC+=$(patsubst $(LIB_DIR)/hardware/boards/$(FPGA_TOOL)/$(BOARD)/%, $(BUILD_DIR)/hw/fpga/%, $(wildcard $(LIB_DIR)/hardware/boards/$(FPGA_TOOL)/$(BOARD)/*))
@@ -151,20 +143,17 @@ $(BUILD_FPGA_DIR)/%: $(FPGA_DIR)/$(FPGA_TOOL)/$(BOARD)/%
 endif
 #--------------------- DOCUMENT-----------------------
 ifneq ($(wildcard document/.),)
+
 SRC+=$(BUILD_DOC_DIR)
-$(BUILD_DOC_DIR): $(LIB_DIR)/document
-	cp -r $< $@
+$(BUILD_DOC_DIR):
+	mkdir -p $@/tsrc
 
-SRC+=$(patsubst $(LIB_DIR)/document/%, $(BUILD_DOC_DIR)/%, $(shell find $(LIB_DIR)/document))
-$(BUILD_DOC_DIR)/%: $(LIB_DIR)/document/%
-	cp $< $@
-
-SRC+=$(BUILD_TSRC_DIR)/shortHash.tex
-$(BUILD_TSRC_DIR)/shortHash.tex:
+SRC+=$(BUILD_DIR)/doc/tsrc/shortHash.tex
+$(BUILD_DIR)/doc/tsrc/shortHash.tex:
 	git rev-parse --short HEAD > $@
 
-SRC+=$(patsubst %, $(BUILD_TSRC_DIR)/%, $(wildcard mkregs.conf))
-$(BUILD_TSRC_DIR)/mkregs.conf: mkregs.conf
+SRC+=$(patsubst %, $(BUILD_DIR)/doc/tsrc/%, $(wildcard mkregs.conf))
+$(BUILD_DIR)/doc/tsrc/mkregs.conf: mkregs.conf
 	cp $< $@
 
 SRC+=$(patsubst $(DOC_DIR)/%, $(BUILD_DOC_DIR)/%, $(wildcard $(DOC_DIR)/*.expected))
@@ -175,21 +164,26 @@ SRC+=$(patsubst $(DOC_DIR)/%, $(BUILD_DOC_DIR)/%, $(wildcard $(DOC_DIR)/document
 $(BUILD_DOC_DIR)/document.mk: $(DOC_DIR)/document.mk
 	cp $< $@
 
-SRC+=$(patsubst $(DOC_DIR)/%, $(BUILD_TSRC_DIR)/%, $(wildcard $(DOC_DIR)/*.tex))
-$(BUILD_TSRC_DIR)/%: $(DOC_DIR)/%.tex
+SRC+=$(patsubst $(DOC_DIR)/%, $(BUILD_DIR)/doc/tsrc/%, $(wildcard $(DOC_DIR)/*.tex))
+$(BUILD_DIR)/doc/tsrc/%.tex: $(DOC_DIR)/%.tex
 	cp $< $@
 
-SRC+=$(patsubst $(DOC_DIR)/figures/%, $(BUILD_FIG_DIR)/%, $(wildcard $(DOC_DIR)/figures/*))
-$(BUILD_FIG_DIR)/%: $(DOC_DIR)/figures/%
+SRC+=$(BUILD_DIR)/doc/figures
+$(BUILD_DIR)/doc/figures: $(DOC_DIR)/figures
+	cp -r $< $@
+
+SRC+=$(BUILD_DIR)/sw/python/verilog2tex.py
+$(BUILD_DIR)/sw/python/verilog2tex.py: $(LIB_DIR)/software/python/verilog2tex.py
 	cp $< $@
 
-SRC+=$(BUILD_SW_PYTHON_DIR)/verilog2tex.py
-$(BUILD_SW_PYTHON_DIR)/verilog2tex.py: $(LIB_DIR)/software/python/verilog2tex.py
+SRC+=$(BUILD_DIR)/sw/python/mkregs.py
+$(BUILD_DIR)/sw/python/mkregs.py: $(LIB_DIR)/software/python/mkregs.py
 	cp $< $@
 
-SRC+=$(BUILD_SW_PYTHON_DIR)/mkregs.py
-$(BUILD_SW_PYTHON_DIR)/mkregs.py: $(LIB_DIR)/software/python/mkregs.py
-	cp $< $@
+SRC+=$(patsubst $(LIB_DIR)/document/%, $(BUILD_DOC_DIR)/%, $(shell find $(LIB_DIR)/document))
+$(BUILD_DOC_DIR)/%: $(LIB_DIR)/document/%
+	if [ ! -f $(DOC_DIR)/`basename $<` ]; then cp $< $@; fi
+
 endif
 
 all: setup
@@ -226,16 +220,15 @@ clean:
 	@rm -rf software/python/__pycache__
 
 debug: $(BUILD_DIR) $(VHDR) 
-	@echo $(NAME)
-	@echo $(TOP_MODULE)
-	@echo $(VERSION)
-	@echo $(VERSION_STR)
-	@echo $(BUILD_DIR)
-	@echo $(BUILD_VSRC_DIR)
-	@echo $(BUILD_SW_SRC_DIR)
-	@echo $(SRC)
-	@echo $(SIMULATOR)
-	@echo $(BOARD)
-	@echo $(FPGA_TOOL)
+	@echo NAME=$(NAME)
+	@echo TOP_MODULE=$(TOP_MODULE)
+	@echo VERSION=$(VERSION)
+	@echo VERSION_STR=$(VERSION_STR)
+	@echo BUILD_DIR=$(BUILD_DIR)
+	@echo BUILD_VSRC_DIR=$(BUILD_VSRC_DIR)
+	@echo SRC=$(SRC)
+	@echo SIMULATOR=$(SIMULATOR)
+	@echo BOARD=$(BOARD)
+	@echo FPGA_TOOL=$(FPGA_TOOL)
 
 .PHONY: all setup clean debug
